Refactor user management pages: remove unused components, implement user grouping by role, and enhance user data structure
